--- a/src/Bandersnatch/Bandersnatch.sln
+++ b/src/Bandersnatch/Bandersnatch.sln
@@ -46,8 +46,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "benchmarks", "benchmarks", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nethermind.Field.Montgomery.Benchmark", "Nethermind.Field.Montgomery.Benchmark\Nethermind.Field.Montgomery.Benchmark.csproj", "{6E8DBA09-DF44-4861-AA9F-E23E435D33F2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nethermind.Field.Arithmetic.Benchmark", "Nethermind.Field.Arithmetic.Benchmark\Nethermind.Field.Arithmetic.Benchmark.csproj", "{D356588B-A2EE-437D-BB73-4BCB0D03A3C5}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -138,10 +136,6 @@ Global
 		{6E8DBA09-DF44-4861-AA9F-E23E435D33F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6E8DBA09-DF44-4861-AA9F-E23E435D33F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6E8DBA09-DF44-4861-AA9F-E23E435D33F2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D356588B-A2EE-437D-BB73-4BCB0D03A3C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D356588B-A2EE-437D-BB73-4BCB0D03A3C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D356588B-A2EE-437D-BB73-4BCB0D03A3C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D356588B-A2EE-437D-BB73-4BCB0D03A3C5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{A2D1D2C4-5990-4B17-9D93-124B0D72373E} = {70B362E0-594F-40D3-A97B-B47335654175}
@@ -152,6 +146,5 @@ Global
 		{BDA8184D-5314-4143-A4CA-B062045ABE66} = {70B362E0-594F-40D3-A97B-B47335654175}
 		{09D6A50C-5573-4B9B-83A5-5D46452F171E} = {70B362E0-594F-40D3-A97B-B47335654175}
 		{6E8DBA09-DF44-4861-AA9F-E23E435D33F2} = {3D07B673-FA14-453E-9E27-BF7742DFBE0E}
-		{D356588B-A2EE-437D-BB73-4BCB0D03A3C5} = {3D07B673-FA14-453E-9E27-BF7742DFBE0E}
 	EndGlobalSection
 EndGlobal