--- a/src/Bandersnatch/Nethermind.Field.Arithmetic.Benchmark/Program.cs
+++ b/src/Bandersnatch/Nethermind.Field.Arithmetic.Benchmark/Program.cs
@@ -1,0 +1,16 @@
+using System.Diagnostics;
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Running;
+
+namespace Nethermind.Field.Arithmetic.Benchmark
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            IConfig config = Debugger.IsAttached ? new DebugInProcessConfig() : DefaultConfig.Instance;
+
+            BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args, config);
+        }
+    }
+}