Before

``` ini

BenchmarkDotNet=v0.13.2, OS=pop 22.04
AMD Ryzen 9 7900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK=7.0.102
  [Host]   : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  .NET 7.0 : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2

Job=.NET 7.0  Runtime=.NET 7.0  

```
|             Method |                  _a |                  _b |      Mean |     Error |    StdDev |    Median | Ratio | RatioSD | Allocated | Alloc Ratio |
|------------------- |-------------------- |-------------------- |----------:|----------:|----------:|----------:|------:|--------:|----------:|------------:|
| **Inverse_BigInteger** | **(115(...)FpE) [204]** | **(115(...)FpE) [204]** | **55.253 μs** | **1.4460 μs** | **4.2635 μs** | **55.054 μs** |  **1.00** |    **0.00** |     **112 B** |        **1.00** |
|    Inverse_UInt256 | (115(...)FpE) [204] | (115(...)FpE) [204] | 44.198 μs | 0.8036 μs | 0.7516 μs | 43.799 μs |  0.83 |    0.05 |         - |        0.00 |
|    Inverse_Element | (115(...)FpE) [204] | (115(...)FpE) [204] |  6.332 μs | 0.0478 μs | 0.0447 μs |  6.328 μs |  0.12 |    0.01 |         - |        0.00 |
|                    |                     |                     |           |           |           |           |       |         |           |             |
| **Inverse_BigInteger** | **(115(...)FpE) [204]** | **(619(...)FpE) [200]** | **59.165 μs** | **1.1757 μs** | **2.9280 μs** | **60.154 μs** |  **1.00** |    **0.00** |     **112 B** |        **1.00** |
|    Inverse_UInt256 | (115(...)FpE) [204] | (619(...)FpE) [200] | 46.084 μs | 0.4095 μs | 0.3831 μs | 46.064 μs |  0.78 |    0.06 |         - |        0.00 |
|    Inverse_Element | (115(...)FpE) [204] | (619(...)FpE) [200] |  6.345 μs | 0.0153 μs | 0.0144 μs |  6.344 μs |  0.11 |    0.01 |         - |        0.00 |
|                    |                     |                     |           |           |           |           |       |         |           |             |
| **Inverse_BigInteger** | **(619(...)FpE) [200]** | **(115(...)FpE) [204]** | **60.682 μs** | **2.5121 μs** | **7.4071 μs** | **62.673 μs** |  **1.00** |    **0.00** |     **112 B** |        **1.00** |
|    Inverse_UInt256 | (619(...)FpE) [200] | (115(...)FpE) [204] | 45.103 μs | 0.3551 μs | 0.3321 μs | 45.001 μs |  0.73 |    0.08 |         - |        0.00 |
|    Inverse_Element | (619(...)FpE) [200] | (115(...)FpE) [204] |  6.466 μs | 0.0362 μs | 0.0302 μs |  6.476 μs |  0.10 |    0.01 |         - |        0.00 |
|                    |                     |                     |           |           |           |           |       |         |           |             |
| **Inverse_BigInteger** | **(619(...)FpE) [200]** | **(619(...)FpE) [200]** | **58.551 μs** | **2.2608 μs** | **6.6661 μs** | **59.547 μs** |  **1.00** |    **0.00** |     **112 B** |        **1.00** |
|    Inverse_UInt256 | (619(...)FpE) [200] | (619(...)FpE) [200] | 46.268 μs | 0.2618 μs | 0.2449 μs | 46.215 μs |  0.84 |    0.09 |         - |        0.00 |
|    Inverse_Element | (619(...)FpE) [200] | (619(...)FpE) [200] |  6.484 μs | 0.0375 μs | 0.0313 μs |  6.486 μs |  0.12 |    0.01 |         - |        0.00 |





After


``` ini

BenchmarkDotNet=v0.13.2, OS=pop 22.04
AMD Ryzen 9 7900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK=7.0.102
  [Host]   : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  .NET 7.0 : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2

Job=.NET 7.0  Runtime=.NET 7.0  

```
|             Method |                  _a |                  _b |      Mean |     Error |    StdDev |    Median | Ratio | RatioSD | Allocated | Alloc Ratio |
|------------------- |-------------------- |-------------------- |----------:|----------:|----------:|----------:|------:|--------:|----------:|------------:|
| **Inverse_BigInteger** | **(115(...)FpE) [204]** | **(115(...)FpE) [204]** | **56.104 μs** | **1.5886 μs** | **4.6840 μs** | **56.782 μs** |  **1.00** |    **0.00** |     **112 B** |        **1.00** |
|    Inverse_UInt256 | (115(...)FpE) [204] | (115(...)FpE) [204] | 46.936 μs | 0.5594 μs | 0.4959 μs | 46.945 μs |  0.79 |    0.05 |         - |        0.00 |
|    Inverse_Element | (115(...)FpE) [204] | (115(...)FpE) [204] |  1.671 μs | 0.0047 μs | 0.0044 μs |  1.669 μs |  0.03 |    0.00 |         - |        0.00 |
|                    |                     |                     |           |           |           |           |       |         |           |             |
| **Inverse_BigInteger** | **(115(...)FpE) [204]** | **(619(...)FpE) [200]** | **56.879 μs** | **1.5038 μs** | **4.4341 μs** | **58.847 μs** |  **1.00** |    **0.00** |     **112 B** |        **1.00** |
|    Inverse_UInt256 | (115(...)FpE) [204] | (619(...)FpE) [200] | 46.508 μs | 0.6175 μs | 0.5474 μs | 46.608 μs |  0.79 |    0.05 |         - |        0.00 |
|    Inverse_Element | (115(...)FpE) [204] | (619(...)FpE) [200] |  1.670 μs | 0.0097 μs | 0.0081 μs |  1.667 μs |  0.03 |    0.00 |         - |        0.00 |
|                    |                     |                     |           |           |           |           |       |         |           |             |
| **Inverse_BigInteger** | **(619(...)FpE) [200]** | **(115(...)FpE) [204]** | **61.227 μs** | **2.3349 μs** | **6.8844 μs** | **64.248 μs** |  **1.00** |    **0.00** |     **112 B** |        **1.00** |
|    Inverse_UInt256 | (619(...)FpE) [200] | (115(...)FpE) [204] | 49.014 μs | 0.9713 μs | 0.9086 μs | 49.184 μs |  0.82 |    0.10 |         - |        0.00 |
|    Inverse_Element | (619(...)FpE) [200] | (115(...)FpE) [204] |  1.881 μs | 0.0068 μs | 0.0064 μs |  1.881 μs |  0.03 |    0.00 |         - |        0.00 |
|                    |                     |                     |           |           |           |           |       |         |           |             |
| **Inverse_BigInteger** | **(619(...)FpE) [200]** | **(619(...)FpE) [200]** | **60.174 μs** | **2.2377 μs** | **6.5979 μs** | **61.170 μs** |  **1.00** |    **0.00** |     **112 B** |        **1.00** |
|    Inverse_UInt256 | (619(...)FpE) [200] | (619(...)FpE) [200] | 46.255 μs | 0.9185 μs | 1.1616 μs | 45.773 μs |  0.76 |    0.07 |         - |        0.00 |
|    Inverse_Element | (619(...)FpE) [200] | (619(...)FpE) [200] |  1.877 μs | 0.0047 μs | 0.0037 μs |  1.877 μs |  0.03 |    0.00 |         - |        0.00 |